requirements.txt
Django==4.2.7
djangorestframework==3.14.0
boto3==1.34.0
python-decouple==3.8
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.9
